Made a single component called BlogsArray and integrated it in all blog related routes
