--- src/Components/drawer/blogs/blogsList/blogs.js
@@ -1,86 +1,0 @@
-const blogs = [
-  {
-    id: 1,
-    name: 'Learn to code, drink coffee, change the world',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-    updated: 'Updated 11 hours ago',
-  },
-  {
-    id: 2,
-    name: 'Learn to code, change the world',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-    updated: 'Updated 11 hours ago',
-  },
-  {
-    id: 3,
-    name: 'Learn to code, drink coffee',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-    updated: 'Updated 11 hours ago',
-  },
-  {
-    id: 4,
-    name: 'Single Round Match #03',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-    updated: 'Updated 11 hours ago',
-  },
-  {
-    id: 5,
-    name: 'Single Round Match #03',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-    updated: 'Updated 11 hours ago',
-  },
-  {
-    id: 6,
-    name: 'Single Round Match #03',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-    updated: 'Updated 11 hours ago',
-  },
-  {
-    id: 7,
-    name: 'Single Round Match #03',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-    updated: 'Updated 11 hours ago',
-  },
-
-  {
-    id: 8,
-    name: 'Single Round Match #03',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-    updated: 'Updated 11 hours ago',
-  },
-  {
-    id: 9,
-    name: 'Single Round Match #03',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-    updated: 'Updated 11 hours ago',
-  },
-];
-
-export default blogs;
