--- src/Components/user/profile/myPosts.js
@@ -1,20 +1,0 @@
-const myPosts = [
-  {
-    id: 1,
-    name: 'Single Round Match #03',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-  },
-  {
-    id: 2,
-    name: 'Single Round Match #03',
-    tags: ['SRM', 'SRM03', 'contest'],
-    author: 'Swapnil Satish Shinde',
-    date: 'Jan 4, 2019',
-    timeToRead: '2 min read',
-  },
-];
-
-export default myPosts;
